fix: in case of selection don´t remove space

+++ src/syncfusion/index.ts
@@ -539,9 +539,10 @@ export class IaraSyncfusionAdapter
 
   private _handleFirstInference(inference: IaraSpeechRecognitionDetail): void {
     this._updateSelectedNavigationField(this._documentEditor.selection.text);
+    const hasSelectedText = this._documentEditor.selection.text.length
 
     if (this._documentEditor.selection.text.length) {
-      this._documentEditor.editor.delete();
+        this._documentEditor.editor.delete();
     }
 
     this._selectionManager = new IaraSyncfusionSelectionManager(
@@ -565,8 +566,10 @@ export class IaraSyncfusionAdapter
       this._selectionManager.moveSelectionToBeforeBookmarkEdge(
         this._selectionManager.initialSelectionData.bookmarkId
       );
-      this._documentEditor.selection.extendBackward();
-      this._documentEditor.editor.delete();
+      if (!hasSelectedText) {
+        this._documentEditor.selection.extendBackward();
+        this._documentEditor.editor.delete();
+      }
       this._selectionManager.resetSelection();
     }
   }
